Implemented Header and renamed the Files
