+++ src/components/Header/header.css
@@ -1,0 +1,20 @@
+.header{
+    max-width: 950px;
+    margin: 30px auto;
+    display: flex;
+    }
+.image{
+    flex: 1;
+}
+
+.nav-link{
+    margin: auto;
+    padding:0 10px;
+    color: black;
+    font-weight:bold;
+    text-decoration: none;
+}
+
+.nav-link:hover{
+    color:coral;
+}
