+++ src/components/Route.js
@@ -1,0 +1,4 @@
+const Route = ({path,children}) =>{
+    return window.location.pathname === path ? children : null;
+}
+export default Route;
