+++ src/components/Videos/Videos.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Sections from "./video_table/table_body";
+import Sections from "./table_body";
 import "./Videos.css"
 const videoData=[
     {
